requirements.txt
# Core dependencies for PII Redaction Demo
agent-framework>=1.0.0rc5
azure-identity>=1.25.0
azure-monitor-opentelemetry>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# PII Detection dependencies
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
azure-ai-textanalytics>=5.4.0

# Additional dependencies that may be needed
# Uncomment if you encounter import errors:
# azure-core>=1.30.0
# opentelemetry-api>=1.20.0
# openai>=1.0.0